Add explicit cast to uint32_t

+++ ff_fat.c
@@ -712,7 +712,7 @@ uint32_t FF_FindEndOfChain( FF_IOManager_t * pxIOManager,
     {
         /* Traverse FAT for (2^32-1) items/clusters,
          * or until end-of-chain is encountered. */
-        ulFatEntry = FF_TraverseFAT( pxIOManager, ulStart, ~0UL, &xError );
+        ulFatEntry = FF_TraverseFAT( pxIOManager, ulStart, ( uint32_t ) ~0UL, &xError );
     }
     else
     {
